refactor: savings-calculator 디렉토리를 pages 하위 디렉토리로 이동
